Refactors areas brought up in meeting on 9/3
