+++ tests/test_unary_ops.py
@@ -1,4 +1,3 @@
-import pytest
 from operations import unary_ops
 
 def test_square_with_quinary_input():
@@ -7,4 +6,6 @@ def test_square_with_quinary_input():
 def test_square_root_with_quinary_input():
     assert unary_ops.square_root("14") == "3"
 
-#TODO: add more unit tests
+def test_square_root_edge_case():
+    assert unary_ops.square_root("0") == "0"
+    assert unary_ops.square_root("1") == "1"
